app/src/main/java/com/wiryadev/snapcoding/ui/settings/SettingsViewModel.kt
package com.wiryadev.snapcoding.ui.settings

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.asLiveData
import androidx.lifecycle.viewModelScope
import com.wiryadev.snapcoding.data.Result
import com.wiryadev.snapcoding.data.repository.auth.AuthRepository
import com.wiryadev.snapcoding.model.User
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.collect
import kotlinx.coroutines.launch
import javax.inject.Inject

@HiltViewModel
class SettingsViewModel @Inject constructor(
    private val authRepository: AuthRepository
) : ViewModel() {

    fun getUser(): LiveData<Result<User>> {
        return authRepository.getLoggedInUser().asLiveData()
    }

    fun logout() {
        viewModelScope.launch {
            authRepository.removeUserSession().collect()
        }
    }

}